wrapper class for ScopeManager
